Limit char quad matching.
QA noticed for a particular document viewing it caused an out of bounds exception.

This is a legacy viewer sample which had a lot of things crammed into it.  In this case text highlighting isn't quite handling unicode correctly but we have dedicated Text samples for showing
how to work with the WordFinder so let's just make sure the text size and # of quads match up to prevent this problem from happening.